Problem: users can submit more applications #214
Solution:
	Added a function that displays a properly configured modal and redirect after it's confirmed
	Added above modal into homepage and project questions
